Moves part of index spec to a show spec.
