+++ app/controllers/admin/merchants_controller.rb
@@ -4,4 +4,8 @@ class Admin::MerchantsController < ApplicationController
     @merchant = User.find(params[:id])
     @top_five_items = current_user.top_items_for_merchant
   end
+
+  def index
+    @merchants = User.where(role: :merchant)
+  end
 end
