+++ app/controllers/merchants_controller.rb
@@ -1,10 +1,5 @@
 class MerchantsController < ApplicationController
   def index
-    if current_admin?
-      @merchants = User.where(role: :merchant)
-    else
       @merchants = User.active_merchants
-    end
   end
-
 end
