+++ app/controllers/users_controller.rb
@@ -1,4 +1,5 @@
 class UsersController < ApplicationController
+
  def new
    @user = User.new
  end
@@ -23,4 +24,5 @@ class UsersController < ApplicationController
    def user_params
      params.require(:user).permit(:name, :email, :address, :city, :state, :zip, :password, :password_confirmation)
    end
+
 end
