+++ app/views/admin/merchants/index.html.erb
@@ -1,0 +1,13 @@
+<div>
+  <%@merchants.each do |merchant| %>
+    <h1>Name:<a href="<%= admin_merchant_path(merchant.id) %>"><%= merchant.name %></a></h1>
+    <p>State: <%= merchant.state %></p>
+    <p>City: <%= merchant.city %></p>
+    <p>Active: <%= merchant.active %></p>
+    <% if merchant.active %>
+      <%= link_to "Disable Merchant" %>
+    <% elsif !merchant.active %>
+      <%= link_to "Enable Merchant" %>
+    <% end %>
+  <% end %>
+</div>
