+++ app/views/merchants/index.html.erb
@@ -6,17 +6,6 @@
         <p>State: <%= merchant.state %></p>
         <p>City: <%= merchant.city %></p>
         <p>Registered: <%= merchant.date_registered %></p>
-      <% elsif current_admin? %>
-        <h1>Name:<a href="<%= admin_merchant_path(merchant.id) %>"><%= merchant.name %></a></h1>
-        <p>State: <%= merchant.state %></p>
-        <p>City: <%= merchant.city %></p>
-        <p>Active: <%= merchant.active %></p>
-        <% if merchant.active %>
-          <%= link_to "Disable Merchant" %>
-        <% elsif !merchant.active %>
-          <%= link_to "Enable Merchant" %>
-        <% end %>
       <% end %>
-    </div>
   <% end %>
 </div>
